Extracted music drawing.
